+++ packages/utils/crypto/src/index.ts
@@ -6,5 +6,5 @@ export * from "./sepc256k1.js";
 export * from "./x25519.js";
 export * from "./encryption.js";
 export * from "./mix.js";
-export * from "./ipfs.js";
+export * from "./libp2p.js";
 export * from "./utils.js";
